backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
motor==3.6.0
pydantic==2.9.2
python-dotenv==1.0.1
bcrypt==4.2.0
cloudinary==1.41.0